refactor: User의 이름, 비밀번호에 대한 유효성 검사를 service에서 domain으로 변경
